adding third version of voltage sensor driver

+++ voltageSensor.h
@@ -8,7 +8,7 @@
 
 // Resolution of the ADC (e.g., 12-bit ADC has 2^12 = 4096 steps)
 #define ADC_RESOLUTION 4096.0
-
+#include <cstdint>
 // Resistor values in the voltage divider circuit (in ohms)
 #define R1             30000.0 // Resistor R1 in the voltage divider
 #define R2             7500.0  // Resistor R2 in the voltage divider
@@ -19,6 +19,6 @@
 //   voltage - A reference to a float variable where the calculated voltage will be stored
 // Returns:
 //   errorState - ES_OK if the voltage is calculated successfully, ES_OUT_OF_RANGE if the pin is invalid
-errorState CalculateVoltage(int pinNumber, float &voltage);
+errorState CalculateVoltage(int8_t pinNumber, float &voltage);
 
 #endif // End of include guard VOLT_H_
